// src/app/models/user.ts
export class User {
  name: string | undefined;
  phone: string | undefined;
  main_address: string | undefined;
  email: string | undefined;
  password: string | undefined;
  date_of_birth: string | undefined;
  id_number: number | undefined;
  job: string | undefined;
  gender: string | undefined;
  photo: File | undefined;
  id_photo: File  | undefined;
}